+++ Sources/VComponents/Components/Modals/Alert/VAlertUIModel.swift
@@ -78,10 +78,22 @@ public struct VAlertUIModel {
 
         uiModel.backgroundColor = backgroundColor
 
+        uiModel.borderWidth = borderWidth
+        uiModel.borderColor = borderColor
+
         uiModel.contentMargins = .zero
 
         return uiModel
     }
+
+    // MARK: Properties - Border
+    /// Border width. Set to `0` points.
+    ///
+    /// To hide border, set to `0`.
+    public var borderWidth: PointPixelMeasurement = .points(0)
+
+    /// Border color.
+    public var borderColor: Color = .clear
 
     // MARK: Properties - Alert Content
     /// Additional margins applied to title text, message text, and content as a whole. Set to `(15, 15, 15, 10)`.